ctrl + s everywhere

+++ Structures/Vector2D.cpp
@@ -7,7 +7,7 @@ Vector2D::Vector2D()
 	y = 0;
 }
 
-Vector2D::Vector2D(float x, float y) {
+Vector2D::Vector2D(int x, int y) {
 	this->x = x;
 	this->y = y;
 }
